+++ wayland-client/examples/dynamic_globals.rs
@@ -14,7 +14,7 @@ fn main() {
 
     let mut event_queue = display.create_event_queue();
 
-    let attached_display = (*display).clone().attach(event_queue.get_token());
+    let attached_display = (*display).clone().attach(event_queue.token());
 
     // We create a GlobalManager with a callback, that will be
     // advertised of any global creation or deletion