+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [client/server] Introduce the `DispatchData` mechanism, allowing global state to be shared with the
   filters and callbaks during an event dispatch.
 - [scanner] Use `#[non_exhaustive]` on protocol-generated enums to replace the `__nonexhaustive` hidden variant
+- [client] Rename `EventQueue::get_token` to `EventQueue::token`
 
 ## 0.24.1 -- 2019-12-13